+++ bin/free.c
@@ -6,17 +6,15 @@
 #include <util.h>
 
 int free() {
-  char msg[128];
   char *fmt =
     "      total  used  free\n"
-    "Mem:  %d      %d     %d\n";
+    "Mem:  %d      %d     %d";
 
   struct MEM_INFO mem = {
     0, 0, 0
   };
   stats(&mem);
 
-  sprintf(msg, fmt, mem.total, mem.used, mem.free);
-  put_str(msg, CHAR_COLOR);
+  printf(fmt, mem.total, mem.used, mem.free);
   return EXIT_SUCCESS;
 }